+++ ValheimEnforcer/ValheimEnforcer.cs
@@ -29,7 +29,7 @@ namespace ValheimEnforcer
     {
         public const string PluginGUID = "MidnightsFX.ValheimEnforcer";
         public const string PluginName = "ValheimEnforcer";
-        public const string PluginVersion = "0.23.2";
+        public const string PluginVersion = "0.24.0";
 
         internal static ManualLogSource Log;
         internal ValConfig cfg;